requirements.txt
asgiref==3.7.2
coverage==7.3.2
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==6.1.0
inflection==0.5.1
mccabe==0.7.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
